Add register_table and deregister_table
