+++ src/catalog.rs
@@ -58,6 +58,10 @@ impl PyTable {
     pub fn new(table: Arc<dyn TableProvider>) -> Self {
         Self { table }
     }
+
+    pub fn table(&self) -> Arc<dyn TableProvider> {
+        self.table.clone()
+    }
 }
 
 #[pymethods]
